feat : 테스트 상황시 예외를 발생시키는 클래스 생성
