+++ src/main/resources/test-applicationContext.xml
@@ -14,7 +14,7 @@
         <constructor-arg ref="dataSource"/>
     </bean>
     <bean id="userLevelUpgradePolicy" class="springbook.user.service.DefaultUpgradePolicy"/>
-    <bean id="userService" class="springbook.user.service.UserService">
+    <bean id="userService" class="springbook.user.service.UserServiceImpl">
         <constructor-arg name="userDao" ref="userDao"/>
         <constructor-arg name="userLevelUpgradePolicy" ref="userLevelUpgradePolicy"/>
     </bean>
