+++ src/test/java/springbook/user/service/TestUserServiceException.java
@@ -1,0 +1,7 @@
+package springbook.user.service;
+
+public class TestUserServiceException extends RuntimeException {
+    public TestUserServiceException(String message) {
+        super(message);
+    }
+}
